all bacis setup

+++ src/main/java/com/bidding/repository/UserRepository.java
@@ -3,9 +3,9 @@ package com.bidding.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.bidding.entity.User;
+import com.bidding.domain.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
 }
